fix(scheme): cast module pressure scores to numpy float64
weekly_1y_causal_v1_31_0_standalone 的 predict 必然失败：build_module_scores 中
`module_abs.idxmax(axis=1)` 在 12 行暖机期全 NaN 数据上抛
`'indices' contains values less than allowed (-9223372036854775808 < -1)`。

判别点是 dtype 而非数据：同族已激活的 cgb_causal_wk_1y_v128 有逐字节相同的
build_module_scores 与 safe_numeric，其 module_abs 形状、全 NaN 行数、各列非空计数
均与 v131 完全相同，但 dtype 是 numpy float64 而 v131 是可空 Float64。可空类型下
全 NaN 行的逐行 argmax 得到 pd.NA，被转成 int64 最小值而非 -1，pandas 的
take(allow_fill=True) 因此拒绝。

在 idxmax 调用点转为 numpy float64，数值不变（NA -> nan），行为与 v128 一致。
已在 ECS 用同一份 DataBridge 数据验证：修复前必然失败（2026-08-22 与 2026-08-15
均失败），修复后 predict 成功产出 direction=1。

注意：Blackbox 的 scheme_version 只由 config.yaml 派生（code_hash 为空串哈希），
本次代码改动不会改变记录的 scheme_version；其可追溯性由 release 的
source_tree_sha256 承担。上游若重新交付该方案，应在其自身版本中修复同一问题。

+++ schemes/weekly_1y_causal_v1_31_0_standalone/delivery/weekly_1y_causal_v1_31_0_standalone.py
@@ -1329,7 +1329,7 @@ def _build_component_19(_components):
             module_cols.append(score_col)
         out['yield_pressure_score'] = out[module_cols].mean(axis=1)
         out['bond_price_score'] = -out['yield_pressure_score']
-        module_abs = out[module_cols].abs()
+        module_abs = out[module_cols].abs().astype('float64')
         out['top_pressure_module_id'] = module_abs.idxmax(axis=1).str.replace('_yield_pressure', '', regex=False)
         out['top_pressure_module'] = out['top_pressure_module_id'].map(MODULE_CN)
         return out
